Split actor commands into multiple files and update to use schema
